Unix cursor support, some gcc warnings, etc
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3608 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/adts.cpp
@@ -16,7 +16,7 @@ namespace test {
 static void test_ringbuf()
 {
 	const size_t N = 49;	// RingBuf capacity
-	const size_t S = 100;	// number of test items
+	const int S = 100;	// number of test items
 
 	// insert and remove immediately
 	{
